wip: move schema to separate file
